+++ conn.go
@@ -3,7 +3,7 @@
 //
 // SPDX-License-Identifier: UPL-1.0 OR Apache-2.0
 
-package goracle
+package godror
 
 /*
 #include <stdlib.h>
@@ -12,6 +12,7 @@ package goracle
 import "C"
 
 import (
+	"bytes"
 	"context"
 	"database/sql"
 	"database/sql/driver"
@@ -31,7 +32,7 @@ const wrapResultset = "--WRAP_RESULTSET--"
 // The maximum capacity is limited to (2^32 / sizeof(dpiData))-1 to remain compatible
 // with 32-bit platforms. The size of a `C.dpiData` is 32 Byte on a 64-bit system, `C.dpiSubscrMessageTable` is 40 bytes.
 // So this is 2^25.
-// See https://github.com/go-goracle/goracle/issues/73#issuecomment-401281714
+// See https://github.com/go-godror/godror/issues/73#issuecomment-401281714
 const maxArraySize = (1<<32)/C.sizeof_dpiSubscrMessageTable - 1
 
 var _ = driver.Conn((*conn)(nil))
@@ -402,9 +403,9 @@ func (c *conn) init() error {
 			return errors.Errorf("getServerVersion: %w", c.getError())
 		}
 		c.Server.set(&v)
-		c.Server.ServerRelease = string(bytesReplaceAll(
+		c.Server.ServerRelease = string(bytes.Replace(
 			((*[maxArraySize]byte)(unsafe.Pointer(release)))[:releaseLen:releaseLen],
-			[]byte{'\n'}, []byte{';', ' '}))
+			[]byte{'\n'}, []byte{';', ' '}, -1))
 	}
 
 	if c.timeZone != nil && (c.timeZone != time.Local || c.tzOffSecs != 0) {