+++ rows.go
@@ -3,7 +3,7 @@
 //
 // SPDX-License-Identifier: UPL-1.0 OR Apache-2.0
 
-package goracle
+package godror
 
 /*
 #include "dpiImpl.h"