+++ stmt.go
@@ -3,13 +3,22 @@
 //
 // SPDX-License-Identifier: UPL-1.0 OR Apache-2.0
 
-package goracle
+package godror
 
 /*
 #include <stdlib.h>
 #include "dpiImpl.h"
 
 const int sizeof_dpiData = sizeof(void);
+
+void godror_setFromString(dpiVar *dv, uint32_t pos, const _GoString_ value) {
+	uint32_t length;
+	length = _GoStringLen(value);
+	if( length == 0 ) {
+		return;
+	}
+	dpiVar_setFromBytes(dv, pos, _GoStringPtr(value), length);
+}
 */
 import "C"
 import (
@@ -20,6 +29,7 @@ import (
 	"io"
 	"reflect"
 	"strconv"
+	"strings"
 	"sync"
 	"time"
 	"unsafe"
@@ -559,10 +569,6 @@ func (st *statement) NumInput() int {
 			return 1
 		}
 		return 0
-	}
-
-	if !go10 {
-		return -1
 	}
 
 	st.Lock()
@@ -2066,3 +2072,41 @@ type Column struct {
 	Scale                     C.int8_t
 	Nullable                  bool
 }
+
+func dpiSetFromString(dv *C.dpiVar, pos C.uint32_t, x string) {
+	C.godror_setFromString(dv, pos, x)
+}
+
+var stringBuilders = stringBuilderPool{
+	p: &sync.Pool{New: func() interface{} { return &strings.Builder{} }},
+}
+
+type stringBuilderPool struct {
+	p *sync.Pool
+}
+
+func (sb stringBuilderPool) Get() *strings.Builder {
+	return sb.p.Get().(*strings.Builder)
+}
+func (sb *stringBuilderPool) Put(b *strings.Builder) {
+	b.Reset()
+	sb.p.Put(b)
+}
+
+/*
+// ResetSession is called while a connection is in the connection
+// pool. No queries will run on this connection until this method returns.
+//
+// If the connection is bad this should return driver.ErrBadConn to prevent
+// the connection from being returned to the connection pool. Any other
+// error will be discarded.
+func (c *conn) ResetSession(ctx context.Context) error {
+	if Log != nil {
+		Log("msg", "ResetSession", "conn", c.dpiConn)
+	}
+	//subCtx, cancel := context.WithTimeout(ctx, 30*time.Second)
+	//err := c.Ping(subCtx)
+	//cancel()
+	return c.Ping(ctx)
+}
+*/